Start refactoring into more separated parts
